fix: 修掉语义核对发现的 3 处与上游不一致
1) InventoryUtils.getContainerInventory / PlayerStateManager.streamInvContent
   用的是 nonEmptyItemCopyStream()，上游是 yarn ContainerComponent.stream()。
   两者不是一回事：
     - yarn stream()          = 全部槽位（空槽给 EMPTY 占位）→ mojmap allItemsCopyStream()
     - yarn streamNonEmpty()  = 只留非空                 → mojmap nonEmptyItemCopyStream()
   用错会让返回的列表丢掉空槽占位、索引不再与槽位一一对应。
   （两人各自独立发现；PlayerStateManager 那处调用方另有 filter，实际无行为差异，
     但方法选择本身是错的，一并改正。）

2) ElytraBot 的 MaceArua follow-predictor-line：
   predictWithExtraTicks 的 ticks 参数本轮已由 int 改成 double，
   调用点漏删 (int) 强转，把 tick 小数部分的线性插值吃掉了。

3) tr.py：getRotationVec 的替换写成了 .getViewVector($1)，
   而 Python re.sub 只认 \1 —— 产物里会留下字面量 getViewVector($1)。
   另补 yarn Entity.getRotationVector() → getLookAngle()
   （26.2 同名方法返回 Vec2，不能映射成同名）。

+++ src/main/java/me/matl114/hacks/modules/combat/ElytraBot.java
@@ -1787,7 +1787,7 @@ public class ElytraBot extends BaseModule {
                         Vec3 predictionPosition = PositionPredict.INSTANCE
                                 .attackPredictArgument
                                 .get()
-                                .predictWithExtraTicks(base.target, (int) predictionTicks);
+                                .predictWithExtraTicks(base.target, predictionTicks);
                         Vec3 untrustedDirection = calculateTargetDirection(predictionPosition);
                         if (untrustedDirection.y < 0 && conditionMovement.test(untrustedDirection)) {
                             movementDirection = untrustedDirection;
